Style Fixed
The issue was with the way in which I was determining the type of the file. The bootstrap .min.css was being interpreted as a min file, as I was using the JS split() method to split the string at the first . and then taking the section after it. Instead I am now getting the index of the last . and then creating a substring of the remaining text (i.e. the filetype)