Successfully passing PUT request to server
Successfully recieving a PUT request on the server, containing a JSON string of data which the user wants to save. Current issue - the address for the 'get'^X request is causing issue, as in Google Chrome you can't use the hostname to send a request in local testing. Trying to resolve now, but currently this means that only the homepage gets populated with data from the remote json file (as they are unable to locate this resource with the same url as the homepage, because these pages exist in a subfolder, so would need a '../' before the address to access it