Fix on Previous Commit
Minor fix where the name was being concatenated onto the path twice, resolved now. Data being passed to and stored in the server. Currently have a problem with saving over the original data (can only save my json file in the main folder, and the load from the html folder, so currently the two never meet. This is a problem with my use of the File System module, so I will revise this tomorrow. For the time being, I am delighted that data is now passing back and forth between the website and the server.